requirements.txt
azure-identity
azure-mgmt-datafactory